requirements.txt
pandas
numpy
scikit-learn
lightgbm
matplotlib
seaborn
fastapi
uvicorn
pydantic
nannyml
mlflow
pyyaml
pytest